refactor: make scripts/generate.py perform required fixups for openapi-python-client lazy imports (#40)
This PR will enable removal of the `enable_lazy_imports` changes from
our `openapi-python-client` fork, by providing the required alterations
to generated code in `scripts/generate.py`.

In v0.12.0 upstream `openapi-python-client` changed the generated code
to use ForwardRef type annotations in various parts of the generated
code. This was a problem for us because our `formatter` module needs to
resolve those types at runtime as it walks the response object, so the
solution we used at the time was to add a new `enable_lazy_imports`
class attribute to `Property` and use its value to determine whether to
change the behaviour of the `Property` hierarchy back to normal
('eager') imports.

While it would be possible to feed the relevant types to
`typing.get_type_hints()` via its `localns` parameter; since we do not
know ahead of time which particular types are required we would have to
import all types from `models`, which is a pretty significant
performance hit for CLI usage.

By performing the relevant fixups in `generate.py`, we avoid that while
bringing us a step closer to switching back to the upstream
`openapi-python-client`.