feat: support x-cli-lookup attribute on request body properties (#68)
This PR adds support for using x-cli-lookup attribute on request body
properties, which currently affects following commands:

- load-balancer create
- load-balancer update
- load-balancer server create
- load-balancer server delete

If the generated CLI option name ends with x-cli-entity.id or its plural
("_id" or "_ids")
that is removed, resulting in the four commands above now using
`--servers` instead of `--server-ids`.

The former `--server-ids` is still accepted to provide compatibility
with previous releases.